Fix bug with odd/even listeners on Read and Delete button
With help from Odin Project Discord, figured out the issue was that
with every submit of a new book, toggleRead and deleteBook were being
called. So an additional event listener was being attached to all
buttons.

This was fixed by moving the function calls to card creation, using
the known IDs for that card, to create a single event listener for
each card button.